修改cleancode
# message auto-generated for no-merge-commit merge:
merge master_fix_cluster into master

修改cleancode

Created-by: insightDeveloper

Match-id-54f84fc18a79243d9b45e0a7990adb288e258f7c

+++ server/msinsight/src/base/BaseModule.cpp
@@ -18,7 +18,7 @@ void BaseModule::OnRequest(std::unique_ptr<Protocol::Request> request)
     int maxThreadNum = MIN_SIZE;
     unsigned int cpuCount = SystemUtil::GetCpuCoreCount();
     if (cpuCount >= MIN_SIZE && cpuCount <= MAX_SIZE) {
-        maxThreadNum = cpuCount;
+        maxThreadNum = static_cast<int>(cpuCount);
     }
     static ThreadPool threadPool(maxThreadNum);
     std::string command = request->command;
