+++ server/src/modules/memory/handler/QueryMemoryComponentHandler.cpp
@@ -156,8 +156,8 @@ void QueryMemoryComponentHandler::SelectResult(Dic::Protocol::MemoryComponentReq
                                                std::vector<MemoryComponentComparison> &fullDiffResult)
 {
     SortResult(request, fullDiffResult);
-    uint64_t pageSize = request.params.pageSize <= 0 ? DEFAULT_PAGE_SIZE : request.params.pageSize;
-    uint64_t currentPage = request.params.currentPage < 1 ? 0 : request.params.currentPage - 1;
+    uint64_t pageSize = request.params.pageSize <= 0 ? DEFAULT_PAGE_SIZE : static_cast<uint64_t>(request.params.pageSize);
+    uint64_t currentPage = request.params.currentPage < 1 ? 0 : static_cast<uint64_t>(request.params.currentPage - 1);
     uint64_t offset = currentPage * pageSize;
     if (offset != 0 && offset >= fullDiffResult.size()) {
         response.componentDiffDetails.clear();
